changed back: wavelength is now again an int

+++ gui/constrained_parameters_widget.hpp
@@ -39,7 +39,7 @@ public:
     double getMagnetic() const;
     double getRatio() const;
     bool   getWavelengthPattern() const;
-    double    getWavelength() const;
+    int    getWavelength() const;
 
     bool getConstrained() const;
     void setAdvancedValue(const double&);
